Installing ingress-nginx in Tiltfile with Helm
Previously we were installing ingress-nginx separately on the command line using kubectl apply -f [url] before running tilt so that the ingress controller would already exist in dev. This was less than ideal for a reproducible dev build so I decided to install it inside the Tiltfile using the helm_resource extension instead.