Set NODE_ENV=production in build script
We have to set NODE_ENV=development for the container, otherwise the Next dev server won't work right, and the app won't have any styling in dev. However, when we're running the build script we have to have NODE_ENV=production, otherwise Next will log a warning and the build will usually fail.